fix: move app from di
