+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <application
         android:allowBackup="true"
-        android:name="com.london.app.di.NovixApplication"
+        android:name="com.london.app.NovixApplication"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
