+++ presentation/src/main/java/com/london/presentation/screen/home/HomeViewModel.kt
@@ -58,8 +58,6 @@ class HomeViewModel(
     override fun onGenreSelect(genre: Genre) {
         if (genre == state.value.selectedGenre) return
         updateState { copy(selectedGenre = genre) }
-        loadUpcomingMovies(
-            categoryId = if (genre == Genre.All) null else genre.id
-        )
+        loadUpcomingMovies(categoryId = if (genre == Genre.All) null else genre.id)
     }
 }
